extract friends stuff from views in separate file resurrect pause.py
